# metal : add FA-vec kernel for head size 64