Useful for some models such as Llama 3.2 and Whisper

```bash
./scripts/compare-commits.sh master gg/metal-fa-vec-64 -m ./models/llama-3.2-1b-instruct/ggml-model-q8_0.gguf -fa 1 -p 0 -d 0,512,1024,8192 -n 32
```

| Model         | Test       |   t/s master |   t/s gg/metal-fa-vec-64 |   Speedup |
|:--------------|:-----------|-------------:|-------------------------:|----------:|
| llama 1B Q8_0 | tg32       |       229.50 |                   269.18 |      1.17 |
| llama 1B Q8_0 | tg32@d512  |       223.47 |                   260.33 |      1.16 |
| llama 1B Q8_0 | tg32@d1024 |       210.76 |                   253.69 |      1.20 |
| llama 1B Q8_0 | tg32@d8192 |       114.52 |                   203.24 |      1.77 |